packages/cli/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "evalflow"
version = "0.1.0"
description = "pytest for LLMs - catch prompt regressions before they reach production"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["llm", "testing", "ai", "quality", "ci", "evaluation", "prompts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "anthropic>=0.28.0",
]

[project.urls]
Homepage = "https://github.com/emartai/evalflow"
Documentation = "https://emartai.github.io/evalflow/"
Repository = "https://github.com/emartai/evalflow"
Issues = "https://github.com/emartai/evalflow/issues"
Changelog = "https://github.com/emartai/evalflow/blob/main/CHANGELOG.md"

[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.7.0"]
dev = [
    "build>=1.0",
    "hatchling>=1.24.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
]

[project.scripts]
evalflow = "evalflow.main:app"

[tool.hatch.build.targets.wheel]
packages = ["evalflow"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = ["--tb=short", "--strict-markers", "-q"]
markers = [
    "slow: marks tests as slow (deselect with '-m not slow')",
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
]

[tool.coverage.run]
source = ["evalflow"]
omit = ["tests/*", "evalflow/output/*"]

[tool.coverage.report]
fail_under = 82
show_missing = true
